refactor: move LockWalletDialog to ui dir
